Added support for Solaris (#94)
Tested on illumos distribution OmniOS Community Edition.

+++ trantor/utils/Date.cc
@@ -363,7 +363,8 @@ Date::Date(unsigned int year,
     tm.tm_min = minute;
     tm.tm_sec = second;
     epoch = mktime(&tm);
-    microSecondsSinceEpoch_ = epoch * MICRO_SECONDS_PRE_SEC + microSecond;
+    microSecondsSinceEpoch_ =
+        static_cast<int64_t>(epoch) * MICRO_SECONDS_PRE_SEC + microSecond;
 }
 
 }  // namespace trantor
